test_update.rs
Clippy; conditional_chain -> swap chained `if` for `match`/`cmp`.

Add tests for lal::update. These tests document the current
behaviour of the command, but I'm not sure if they are the
behaviours that we want this command to have. As noted in
the comments, some of these behaviours are inconsistent.

e.g. not saving when --save is passed. or not installing a
devDependency when --dev is passed.

Oddly, it is also possible to depend on the same component,
but at different versions (one in `dependencies` and the
other in `devDependencies`).